+++ lib/ruby/1.9/io/console/bsd_console.rb
