+++ lib/ruby/1.9/io/console/linux_console.rb
